Bump to 20.2.1 and mitigate GCC/Linux warnings
Bump DMS_VERSION_PATCH to 1 (20.2.1) and cut the GCC -Wall -Wextra warning
surface on the Linux/CMake build from ~67k to ~450, via source fixes plus a
tuned warning policy in CMakeLists.txt. MSVC (msbuild) and GCC (CMake) both
compile clean (0 errors); no test regressions.

Warning policy (CMakeLists.txt): drop -Wpedantic; -Wno-* the high-noise/low-
signal categories (unused-parameter, deprecated-copy, reorder, unknown-pragmas,
switch, sign-compare, endif-labels, unused-{variable,but-set-variable,local-
typedefs}, deprecated-declarations, overflow).

Largest source fix: name-hiding in the recursive visitor folds — add
`using Base::Visit;` to UnitProcessor.h/ValueClass.h (~9.6k -Woverloaded-virtual).

strict-aliasing: std::bit_cast / static_cast round-trips replacing enum/flag
reference-punning and Float<->UInt reinterprets (Undefined.h, OperArgPolicy.h,
OperAttrBin/Ter.h, FormattedStream.cpp, ThemeSet/GraphicLayer/etc.).

Real bugs fixed:
- OdbcImp: delete[] of void* (UB); %hu printing a ULONG (16-bit truncation).
- XdbImp: fscanf "%d" into a long; removed dead unsigned `< 0` clamp.
- gdal_base: char* == "GML" pointer compare -> string_view; dropped a dead
  enum/int ternary in the read-only branch.
- 1 << (8*sizeof-1) shift-count-overflow (OdbcImp).
- ActorEnums IsInTrans() was always-false; Actor.cpp unreachable case removed.

Behavior-preserving cleanups: ~70 parentheses (&& within ||, assignments),
dangling-else braces, removed dead null-this guards (assert(this)/if(!this)),
dropped ignored top-level const on by-value returns, std::iterator -> explicit
iterator-trait typedefs, MSVC #pragma warning guarded by _MSC_VER, localized
GCC pragmas for the flexible-array-member and bit_value array-bounds idioms,
PartitionTypes Range alias renamed to RangeType, [=,this] lambda captures.

+++ clc/dll/src/random.cpp
@@ -213,7 +213,7 @@ public:
 		parallel_tileloop(te, [result, seed, range](tile_id t)
 			{
 				using uniform_engine_t = uniform_engine<T> ;
-				std::array<UInt32, 2> seedArr = { seed, static_cast<UInt32>(t) };
+				std::array<UInt32, 2> seedArr = { static_cast<UInt32>(seed), static_cast<UInt32>(t) };
 				auto seeds = std::seed_seq(seedArr.begin(), seedArr.end());
 				uniform_engine_t rndEngine(seeds, range.first, range.second); // let each tile have its own seed
